WordPress custom theme compatibility

+++ config/presets.js
@@ -1,11 +1,17 @@
 const fse = require('fs-extra');
 const environment = process.env.NODE_ENV;
 
-const mode = environment === 'react' ? 'react' : 'vue';
-fse.remove('./src/js/', err => {
-  if (err) return console.error(err);
-  fse.copy(`./config/${mode}-config/`, './src/js/', err => {
+if (environment === 'react' || environment === 'vue') {
+  fse.remove('./src/js/', err => {
     if (err) return console.error(err);
-    console.log('Happy hacking!');
+    fse.copy(`./config/${environment}-config/`, './src/js/', err => {
+      if (err) return console.error(err);
+      console.log(`Environment is set to use ${environment} js!`);
+    });
   });
-});
+} else {
+  fse.copy(`./config/wp-config/`, '.', err => {
+    if (err) return console.error(err);
+    console.log(`OtomteJS ❤️  WP!`);
+  });
+}
